change return to double to handle large numbers

+++ 0x08-recursion/3-factorial.c
@@ -4,7 +4,7 @@
  * Return: int
  */
 
-float factorial(int n)
+double factorial(int n)
 {
 	if (n == 0)
 		return (1);
